console class added
