+++ src/main/java/app/Main.java
@@ -1,11 +1,10 @@
-import commands.*;
+package app;
+
 import data.Ticket;
 import parsing.Parser;
 
 import java.io.*;
-import java.nio.file.Path;
 import java.util.*;
-import java.util.regex.Pattern;
 
 /*
  * Это моя 5-ая лабораторная по предмету программирование
@@ -34,6 +33,7 @@ public class Main {
 
         var receiver = new Receiver(storage,filename);
         var invoker = new Invoker(receiver);
-        //invoker.runApp(new InputStreamReader(System.in));
+        var console = new AppConsole();
+        console.runApp(new InputStreamReader(System.in), invoker);
     }
 }
