+++ src/main/java/commands/RemoveById.java
@@ -1,5 +1,7 @@
 package commands;
 
+import app.Receiver;
+
 import java.util.List;
 
 public class RemoveById implements Command{
